+++ src/main/java/io/brachu/johann/cli/ProcessOutputSink.java
@@ -1,0 +1,15 @@
+package io.brachu.johann.cli;
+
+import java.io.IOException;
+
+interface ProcessOutputSink {
+
+    void takeLine(String line);
+
+    void takeErrorLine(String line);
+
+    String standardOutput() throws IOException;
+
+    String errorOutput() throws IOException;
+
+}