+++ src/main/java/io/brachu/johann/cli/ProcessOutputSinkFactory.java
@@ -1,0 +1,8 @@
+package io.brachu.johann.cli;
+
+@FunctionalInterface
+interface ProcessOutputSinkFactory {
+
+    ProcessOutputSink create(Process process);
+
+}