+++ src/main/java/io/brachu/johann/cli/ProcessWaitStrategy.java
@@ -1,0 +1,10 @@
+package io.brachu.johann.cli;
+
+import java.util.concurrent.TimeoutException;
+
+@FunctionalInterface
+interface ProcessWaitStrategy {
+
+    int waitFor(Process process) throws InterruptedException, TimeoutException;
+
+}