+++ src/main/java/io/brachu/johann/cli/SystemProcessOutputSink.java
@@ -1,0 +1,16 @@
+package io.brachu.johann.cli;
+
+final class SystemProcessOutputSink extends PrintStreamProcessOutputSink {
+
+    private static final SystemProcessOutputSink INSTANCE = new SystemProcessOutputSink();
+
+    private SystemProcessOutputSink() {
+        super(System.out, System.err);
+    }
+
+    static SystemProcessOutputSink create(Process process) {
+        ProcessOutputTransfer.start(process, INSTANCE);
+        return INSTANCE;
+    }
+
+}