+++ src/main/java/io/brachu/johann/cli/exception/ExecutionTimedOutException.java
@@ -1,7 +1,0 @@
-package io.brachu.johann.cli.exception;
-
-public class ExecutionTimedOutException extends Exception {
-
-    private static final long serialVersionUID = -4652175280466948956L;
-
-}